refactor: move Antibody to reagents
