+++ src/aind_data_schema/core/procedures.py
@@ -25,9 +25,9 @@ from typing_extensions import Annotated
 
 from aind_data_schema.base import AwareDatetimeWithDefault, DataCoreModel, DataModel
 from aind_data_schema.components.coordinates import Coordinate, CoordinateSystem, Origin
-from aind_data_schema.components.devices import FiberProbe, MyomatrixArray, MyomatrixThread
+from aind_data_schema.components.devices import FiberProbe, MyomatrixArray
 from aind_data_schema.components.identifiers import Person
-from aind_data_schema.components.reagent import Reagent, OligoProbe, HCRProbe, Stain, Fluorophore
+from aind_data_schema.components.reagent import Reagent, OligoProbe, HCRProbe, Stain, Antibody
 from aind_data_schema.utils.merge import merge_notes
 from aind_data_schema.utils.validators import subject_specimen_id_compatibility
 
@@ -156,16 +156,6 @@ class HCRSeries(DataModel):
     hcr_rounds: List[HybridizationChainReaction] = Field(..., title="Hybridization Chain Reaction rounds")
     strip_qc_compatible: bool = Field(..., title="Strip QC compatible")
     cell_id: Optional[str] = Field(default=None, title="Cell ID")
-
-
-class Antibody(Reagent):
-    """Description of an antibody used in immunolableing"""
-
-    immunolabel_class: ImmunolabelClass = Field(..., title="Immunolabel class")
-    fluorophore: Optional[Fluorophore] = Field(default=None, title="Fluorophore")
-    mass: Decimal = Field(..., title="Mass of antibody")
-    mass_unit: MassUnit = Field(default=MassUnit.UG, title="Mass unit")
-    notes: Optional[str] = Field(default=None, title="Notes")
 
 
 class Sectioning(DataModel):
